Changes for JWT Extended

+++ resources/item.py
@@ -1,31 +1,31 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity
 from models.item import ItemModel
+
 
 class Item(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument('price',
                         type=float,
                         required=True,
-                        help="This field cannot be left blank"
+                        help="This field cannot be left blank!"
                         )
     parser.add_argument('store_id',
                         type=int,
                         required=True,
-                        help="Every Item needs a store ID"
+                        help="Every item needs a store_id."
                         )
-#GET Method
-    @jwt_required()
-    def get(self,name):
+
+    @jwt_required  # No longer needs brackets
+    def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
-        return {'message':'Item not found'}, 404
+        return {'message': 'Item not found'}, 404
 
-#POST Method
-    def post(self,name):
+    def post(self, name):
         if ItemModel.find_by_name(name):
-            return {'message': 'A item with name {} already exists'.format(name)}, 400
+            return {'message': "An item with name '{}' already exists.".format(name)}, 400
 
         data = Item.parser.parse_args()
 
@@ -34,24 +34,28 @@ class Item(Resource):
         try:
             item.save_to_db()
         except:
-            return {'message':'An error Occured'}, 500  #Internal Server Error
-        return item.json(),201
+            return {"message": "An error occurred inserting the item."}, 500
 
-#DELETE Method
+        return item.json(), 201
 
-    def delete(self,name):
+    @jwt_required
+    def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message':'Admin privilage required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-            return {'message':'Item Deleted'}
-        return  {'message':'Item not found'}
+            return {'message': 'Item deleted.'}
+        return {'message': 'Item not found.'}, 404
 
-    def put(self,name):
+    def put(self, name):
         data = Item.parser.parse_args()
 
         item = ItemModel.find_by_name(name)
 
-        if item :
+        if item:
             item.price = data['price']
         else:
             item = ItemModel(name, **data)
@@ -62,5 +66,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return{'items':items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you login'
+        }, 200
